+++ CHANGELOG.md
@@ -5,6 +5,16 @@ This document records all notable changes to [nessus file analyzer by LimberDuck
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2021-08-23
+
+### Added
+
+- new columns added to Vulnerabilities report (read more in [documentation](https://nessus-file-analyzer.readthedocs.io/en/latest/nfa-target-file/section-vulnerabilities.html)):
+  -  Exploit available,
+  -  Exploit code maturity,
+  -  Exploit framework metasploit,
+  -  Exploitability ease,
+
 ## [0.5.1] - 2021-03-27
 
 ### Changed
@@ -81,5 +91,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.3.1]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.3.0...v0.3.1
 [0.4.0]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.3.1...v0.4.0
 [0.5.0]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.4.0...v0.5.0
+[0.5.1]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.5.0...v0.5.1
+[0.6.0]: https://github.com/LimberDuck/nessus-file-analyzer/compare/v0.5.1...v0.6.0
 
 [1]: https://github.com/LimberDuck/nessus-file-analyzer